Move stats funcitons into separate file.
